Arguments to the varargs function mps_pool_create must be cast, otherwise the default c promotion rules will result in a constant value lke 64 being protomoted to int (4 bytes) but decoded (wrongly) as size_t (8 bytes).
Copied from Perforce
 Change: 187943
 ServerID: perforce.ravenbrook.com

+++ mps/test/function/224.c
@@ -32,7 +32,7 @@ static void test(void)
  die(mps_arena_commit_limit_set(arena, VMSIZE), "commit limit");
 
  die(mps_pool_create(&pool, arena, mps_class_mv(),
-                     EXTENDBY, AVGSIZE, EXTENDBY),
+                     (size_t)EXTENDBY, (size_t)AVGSIZE, (size_t)EXTENDBY),
      "pool create");
 
  for (p=0; p<ITERATE; p++) {
